flaskapp/pyproject.toml
[project]
name = "sales-dashboard-flask"
version = "0.1.0"
description = "Flask + React dashboard backend for Databricks Apps"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gunicorn>=21.2.0",
    "databricks-sdk>=0.103.0",
    "databricks-sql-connector>=4.2.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[tool.uv]
# This is an app, not a publishable package — install deps only, no local build.
package = false

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["backend/tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
